Migrate PATH setting to .profile
Move PATH setting to .profile to let the login shells take care of
setting environment PATHs once rather than every time a terminal window
is opened.

As .bash_profile is only sourced by bash when started in login mode and
when a user logs in graphically the .profile is sourced during the GNOME
session the path setting is moved to .profile instead of .bash_profile.

c.f.
* https://apple.stackexchange.com/questions/51036/what-is-the-difference-between-bash-profile-and-bashrc
* https://askubuntu.com/questions/121073/why-bash-profile-is-not-getting-sourced-when-opening-a-terminal