Enable mounting kerberos credentials from outside container
If the user has already authenticated with kerberos outside the
container with kinit, then there will be a file /tmp/krb5cc_<uid> where
<uid> is the user's id. As the user inside the container is root and so
has id of 0, then using the user's current `id -u $USER`, the file can
be found outside the container and bind mounted to the /tmp/krb5cc_0
inside the container. This then allows for kerberos protected actions
like xrdcp to be run inside the container.