Add pipx shim
To avoid the complications described in https://stackoverflow.com/q/72357391/8931942
with trying to get pyenv to work nicely with pipx, instead create a
pyenv-virtualenv virtual environment named 'pipx' and then place
the shim file added in this commit at '~/bin/pipx' and make it
executable (chmod +x ~/bin/pipx).

Thanks to Angus Hollands for the idea.